Transfer some cases from test_connection to test_utility
