+++ tests/python_client/testcases/test_utility.py
@@ -2,6 +2,7 @@ import threading
 import time
 
 import pytest
+from pymilvus import DefaultConfig
 from pymilvus.exceptions import MilvusException
 from base.client_base import TestcaseBase
 from base.utility_wrapper import ApiUtilityWrapper
@@ -1711,3 +1712,227 @@ class TestUtilityAdvanced(TestcaseBase):
         self.utility_wrap.load_balance(collection_w.name, src_node_id, dst_node_ids, sealed_segment_ids,
                                        check_task=CheckTasks.err_res,
                                        check_items={ct.err_code: 1, ct.err_msg: "must be in the same replica group"})
+
+
+class TestUtilityUserPassword(TestcaseBase):
+    """ Test case of user interface """
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    def test_create_user_with_user_password(self,host, port):
+        """
+        target: test the user creation with user and password
+        method: create user with the default user and password parameter
+        expected: connected is True
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        user = "nico"
+        password = "wertyu567"
+        self.utility_wrap.create_user(user=user, password=password)
+        self.connection_wrap.connect(host=host, port=port, user=user, password=password,
+                                     check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.list_collections()
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("old_password", ["abc1234"])
+    @pytest.mark.parametrize("new_password", ["abc12345"])
+    def test_reset_password_with_user_and_old_password(self, host, port, old_password, new_password):
+        """
+        target: test the password reset with old password
+        method: get a connection with user and corresponding old password
+        expected: connected is True
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        user = "robot2048"
+        self.utility_wrap.create_user(user=user, password=old_password)
+        self.utility_wrap.reset_password(user=user, old_password=old_password, new_password=new_password)
+        self.connection_wrap.connect(host=host, port=port, user=user,
+                                     password=new_password, check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.list_collections()
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    def test_list_usernames(self, host, port):
+        """
+        target: test the user list created successfully
+        method: get a list of users
+        expected: list all users
+        """
+        #1. default user login
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+
+        #2. create 2 users
+        self.utility_wrap.create_user(user="user1", password="abc123")
+        self.utility_wrap.create_user(user="user2", password="abc123")
+
+        #3. list all users
+        res = self.utility_wrap.list_usernames()[0]
+        assert "user1" and "user2" in res
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("connect_name", [DefaultConfig.DEFAULT_USING])
+    def test_delete_user_with_username(self,host, port,connect_name):
+        """
+        target: test deleting user with username
+        method: delete user with username and connect with the wrong user then list collections
+        expected: deleted successfully
+        """
+        user = "xiaoai"
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.create_user(user=user, password="abc123")
+        self.utility_wrap.delete_user(user=user)
+        self.connection_wrap.disconnect(alias=connect_name)
+        self.connection_wrap.connect(host=host, port=port, user=user,
+                                     password="abc123", check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.list_collections(check_task=ct.CheckTasks.err_res,
+                                           check_items={ct.err_code: 1})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    def test_delete_user_with_invalid_username(self, host, port):
+        """
+        target: test the nonexistant user when deleting credential
+        method: delete a credential with user wrong
+        excepted: delete is true
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.delete_user(user="asdfghj")
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    def test_delete_all_users(self, host, port):
+        """
+        target: delete the users that created for test
+        method: delete the users in list_usernames except root
+        excepted: delete is true
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        res = self.utility_wrap.list_usernames()[0]
+        for user in res:
+            if user != "root":
+                self.utility_wrap.delete_user(user=user)
+        res = self.utility_wrap.list_usernames()[0]
+        assert len(res) == 1
+
+
+class TestUtilityInvalidUserPassword(TestcaseBase):
+    """ Test invalid case of user interface """
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("user", ["qwertyuiopasdfghjklzxcvbnmqwertyui", "@*-.-*", "alisd/"])
+    def test_create_user_with_invalid_username(self, host, port, user):
+        """
+        target: test the user when create user
+        method: make the length of user beyond standard
+        excepted: the creation is false
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.create_user(user=user, password=ct.default_password,
+                                      check_task=ct.CheckTasks.err_res,
+                                      check_items={ct.err_code: 5})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("user", ["alice123w"])
+    def test_create_user_with_existed_username(self, host, port, user):
+        """
+        target: test the user when create user
+        method: create a user, and then create a user with the same username
+        excepted: the creation is false
+        """
+        # 1.default user login
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+
+        # 2.create the first user successfully
+        self.utility_wrap.create_user(user=user, password=ct.default_password)
+
+        # 3.create the second user with the same username
+        self.utility_wrap.create_user(user=user, password=ct.default_password,
+                                      check_task=ct.CheckTasks.err_res, check_items={ct.err_code: 29})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("password", ["12345"])
+    def test_create_user_with_invalid_password(self, host, port, password):
+        """
+        target: test the password when create user
+        method: make the length of user exceed the limitation [6, 256]
+        excepted: the creation is false
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        user = "alice"
+        self.utility_wrap.create_user(user=user, password=password,
+                                      check_task=ct.CheckTasks.err_res, check_items={ct.err_code: 5})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("user", ["hobo89"])
+    @pytest.mark.parametrize("old_password", ["qwaszx0"])
+    def test_reset_password_with_invalid_username(self, host, port, user, old_password):
+        """
+        target: test the wrong user when resetting password
+        method: create a user, and then reset the password with wrong username
+        excepted: reset is false
+        """
+        # 1.default user login
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+
+        # 2.create a user
+        self.utility_wrap.create_user(user=user, password=old_password)
+
+        # 3.reset password with the wrong username
+        self.utility_wrap.reset_password(user="hobo", old_password=old_password, new_password="qwaszx1",
+                                         check_task=ct.CheckTasks.err_res,
+                                         check_items={ct.err_code: 30})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("user", ["demo"])
+    @pytest.mark.parametrize("old_password", ["qwaszx0"])
+    @pytest.mark.parametrize("new_password", ["12345"])
+    def test_reset_password_with_invalid_new_password(self, host, port, user, old_password, new_password):
+        """
+        target: test the new password when resetting password
+        method: create a user, and then set a wrong new password
+        excepted: reset is false
+        """
+        # 1.default user login
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+
+        # 2.create a user
+        self.utility_wrap.create_user(user=user, password=old_password)
+
+        # 3.reset password with the wrong new password
+        self.utility_wrap.reset_password(user=user, old_password=old_password, new_password=new_password,
+                                         check_task=ct.CheckTasks.err_res,
+                                         check_items={ct.err_code: 5})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    @pytest.mark.parametrize("user", ["genny"])
+    def test_reset_password_with_invalid_old_password(self, host, port, user):
+        """
+        target: test the old password when resetting password
+        method: create a credential, and then reset with a wrong old password
+        excepted: reset is false
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.create_user(user=user, password="qwaszx0")
+        self.utility_wrap.reset_password(user=user, old_password="waszx0", new_password="123456",
+                                         check_task=ct.CheckTasks.err_res,
+                                         check_items={ct.err_code: 30})
+
+    @pytest.mark.tags(ct.CaseLabel.L3)
+    def test_delete_user_root(self, host, port):
+        """
+        target: test deleting user root when deleting credential
+        method: connect and then delete the user root
+        excepted: delete is false
+        """
+        self.connection_wrap.connect(host=host, port=port, user=ct.default_user,
+                                     password=ct.default_password, check_task=ct.CheckTasks.ccr)
+        self.utility_wrap.delete_user(user=ct.default_user, check_task=ct.CheckTasks.err_res,
+                                      check_items={ct.err_code: 31})
